fix: correcciones adicionales de deprecaciones PHP 8.5

+++ library/Zend/Locale/Math/PhpMath.php
@@ -198,7 +198,7 @@ class Zend_Locale_Math_PhpMath extends Zend_Locale_Math
             return NULL;
         }
         $result = $op1 % $op2;
-        if (is_nan($result)  or  (($op1 - $result) % $op2 != 0)) {
+        if (is_nan($result)  or  (((int) ($op1 - $result)) % $op2 != 0)) {
             require_once 'Zend/Locale/Math/Exception.php';
             throw new Zend_Locale_Math_Exception("modulus calculation error: $op1 % $op2 != $result", $op1, $op2, $result);
         }
