+++ library/Zend/Validate/Ccnum.php
@@ -97,12 +97,12 @@ class Zend_Validate_Ccnum extends Zend_Validate_Abstract
         $weight = 2;
 
         for ($i = $length - 2; $i >= 0; $i--) {
-            $digit = $weight * $valueFiltered[$i];
+            $digit = (int) ($weight * $valueFiltered[$i]);
             $sum += floor($digit / 10) + $digit % 10;
             $weight = $weight % 2 + 1;
         }
 
-        if ((10 - $sum % 10) % 10 != $valueFiltered[$length - 1]) {
+        if ((10 - (int) $sum % 10) % 10 != $valueFiltered[$length - 1]) {
             $this->_error(self::CHECKSUM, $valueFiltered);
             return false;
         }
